Add files via upload
Based on the RoDi API made in Python. What I did was create as a "terminal" in Python as well. To control RoDI in a slightly easier and more friendly way without writing code.
It is in the process of improvement. I have not finished all the functions, I wanted to boost it so they can see and help if they like.
It is open to suggestions and opinions.